refactor: frontend conversion history
